Reorganize update code
Now that there are multiple update types, reorganize the Python
source files into a better module hierarchy.
